hw1_r13922154/requirements.txt
accelerate==1.10.1
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anywidget==0.9.18
asttokens==3.0.0
bbox_visualizer==0.2.2
certifi==2025.8.3
charset-normalizer==3.4.3
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.4
decorator==5.2.1
defusedxml==0.7.1
einops==0.8.1
ensemble-boxes==1.0.9
executing==2.2.1
fairscale==0.4.13
filelock==3.13.1
filetype==1.2.0
fonttools==4.60.0
fsspec==2024.6.1
ftfy==6.3.1
hf-xet==1.1.10
huggingface-hub==0.35.1
idna==3.7
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.2
jupyter_bbox_widget==0.6.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
llvmlite==0.45.1
MarkupSafe==2.1.5
matplotlib==3.10.6
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.3
ninja==1.13.0
numba==0.62.1
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
open_clip_torch==3.2.0
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.3.2
parso==0.8.5
peft==0.17.1
pexpect==4.9.0
pi_heif==1.1.0
pillow==11.3.0
pillow-avif-plugin==1.5.2
polars==1.34.0
polars-runtime-32==1.34.0
polygraphy==0.49.26
prompt_toolkit==3.0.52
psutil==7.1.0
psygnal==0.14.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycocotools==2.0.10
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pylabel==0.1.55
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rf100vl==1.1.0
rfdetr==1.2.1
roboflow==1.2.9
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==70.2.0
simsimd==6.5.3
six==1.17.0
stack-data==0.6.3
stringzilla==4.0.15
supervision==0.26.1
sympy==1.13.1
threadpoolctl==3.6.0
timm==1.0.20
tokenizers==0.22.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.2
triton==3.1.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.204
ultralytics-thop==2.0.17
urllib3==2.5.0
wcwidth==0.2.14
widgetsnbextension==4.0.14
